Initial Commit
This will take a string of 40 hexadecimal characters check for correct syntax and 5 bit parity. The purpose of this code is to determine if a "credit card" is valid or not. Of course this won't actually be able to pick out real credit cards as it is conceptual and created as a challenge.